feat: separate enum class, edit converter
