+++ src/main/java/study/movie/converter/movie/FilmFormatConverter.java
@@ -1,6 +1,6 @@
 package study.movie.converter.movie;
 
-import study.movie.global.constants.EntityAttrConst.FilmFormat;
+import study.movie.domain.movie.FilmFormat;
 import study.movie.global.converter.EnumArrayConverter;
 
 import javax.persistence.Converter;
