+++ src/main/java/study/movie/converter/movie/FilmRatingConverter.java
@@ -1,10 +1,9 @@
 package study.movie.converter.movie;
 
+import study.movie.domain.movie.FilmRating;
 import study.movie.global.converter.CodeValueConverter;
 
 import javax.persistence.Converter;
-
-import static study.movie.global.constants.EntityAttrConst.*;
 
 @Converter(autoApply = true)
 public class FilmRatingConverter extends CodeValueConverter<FilmRating> {
